backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-ckeditor==6.7.3
django-js-asset==3.1.2
mysqlclient==2.2.7
pillow==11.2.1
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
